requirements.txt
streamlit
beautifulsoup4
requests
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
unstructured
openpyxl
pypdf
msoffcrypto-tool
xlrd>=2.0.1
lxml
tqdm
transformers
torch
Pillow